requirements.txt
streamlit==1.29.0
playwright==1.40.0
pandas==2.1.4
twilio==9.0.0
python-dotenv==1.0.0
alembic==1.13.3
sqlalchemy==2.0.35
psycopg2-binary>=2.9.6
google-generativeai
selenium==4.15.2
webdriver-manager==4.0.2
opencv-python==4.10.0.84
numpy==1.26.4
schedule==1.2.2
tensorflow==2.15.0
beautifulsoup4==4.12.2
pillow==10.3.0
plotly
